+++ SDK.Core/Builders/ActionBuilder.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using MoBro.Plugin.SDK.Enums;
-using MoBro.Plugin.SDK.Models;
+using MoBro.Plugin.SDK.Models.Categories;
 using MoBro.Plugin.SDK.Models.Settings;
 using MoBro.Plugin.SDK.Services;
 using Action = System.Action;