+++ SDK.Core/Models/Categories/Category.cs
@@ -1,24 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace MoBro.Plugin.SDK.Models;
+namespace MoBro.Plugin.SDK.Models.Categories;
 
 /// <summary>
-/// Groups items of a common context (E.g. all metrics of the primary GPU) 
+/// Categorizes items of the same kind (E.g. all CPU related metrics)
 /// </summary>
-/// <remarks>
-/// A group may contain items from different categories
-/// </remarks>
-public sealed class Group : IMoBroItem 
+public sealed class Category : IMoBroItem
 {
   /// <summary>
-  /// Creates a new group with the given properties
+  /// Creates a new category with the given properties
   /// </summary>
-  /// <param name="id">The group id</param>
-  /// <param name="label">The group label</param>
+  /// <param name="id">The category id</param>
+  /// <param name="label">The category label</param>
   /// <param name="description">The optional description</param>
   /// <param name="icon">The optional icon id</param>
-  public Group(string id, string label, string? description = null, string? icon = null)
+  public Category(string id, string label, string? description = null, string? icon = null)
   {
     Id = id ?? throw new ArgumentNullException(nameof(id));
     Label = label ?? throw new ArgumentNullException(nameof(label));
@@ -40,12 +37,12 @@ public sealed class Group : IMoBroItem
   public string? Description { get; set; }
 
   /// <summary>
-  /// An optional icon (relative path to the icon file)
+  /// An optional icon id
   /// </summary>
   public string? Icon { get; set; }
 
   /// <summary>
-  /// Optional sub groups
+  /// Optional sub categories
   /// </summary>
-  public IEnumerable<Group>? SubGroups { get; set; } = new List<Group>();
+  public IEnumerable<Category>? SubCategories { get; set; } = new List<Category>();
 }